[JRO] Better UTF-8 handling
Re: issue #72

Dropped into pry to play around with this a bit and found a manner to
force encoding and handling to be a little less ... delete'y /
text-stripping'y?

Tests now check for both a byte/octet and an actual UTF-8 character
within the context of both html and plain text. Looking at the content
of an email object here, ...

https://gist.github.com/dorra/6354910/raw/1dfa16fd7ad1d70a4286b15438c0472760b6c70d/sample_mail

... thanks to @dorra, provided some context as to where the improvement
could come.